Desarrollo de Tabla LR
Se crearon las constantes donde se definirá el patron de los terminales
y no terminales y también se inicio la generación de la tabla del LR
canónico
