+++ LR1.userprefs
@@ -1,8 +1,9 @@
-﻿<Properties>
+﻿<Properties StartupItem="LR1/LR1.csproj">
   <MonoDevelop.Ide.Workspace ActiveConfiguration="Debug|x86" />
-  <MonoDevelop.Ide.Workbench ActiveDocument="LR1/Program.cs">
+  <MonoDevelop.Ide.Workbench ActiveDocument="LR1/Tools.cs">
     <Files>
-      <File FileName="LR1/Program.cs" Line="7" Column="7" />
+      <File FileName="LR1/Program.cs" Line="3" Column="3" />
+      <File FileName="LR1/Tools.cs" Line="7" Column="7" />
     </Files>
   </MonoDevelop.Ide.Workbench>
   <MonoDevelop.Ide.DebuggingService.Breakpoints>
